fastapi-microservices/injection-service/requirements-ml.txt
faiss-cpu==1.7.4
sentence-transformers==2.2.2
